feat(vue): export `BaseAlertsProps` type definition
Renamed the internal constant `PROPS` to `BASE_ALERTS_PROPS` and updated the type alias from `Props` to `BaseAlertsProps` to clarify the usage and scope within the `BaseAlerts` component. This change also includes updating the component function and properties to use the new `BaseAlertsProps` type, enhancing maintainability and readability of the codebase. This update is crucial for developers extending or integrating with the `BaseAlerts` component, as it makes the props definition publicly accessible and easily importable.