fix(eslint-config): disable `sonarjs/prefer-immediate-return` rule
Disabling the `sonarjs/prefer-immediate-return` rule in the ESLint configuration for the `sonarjs` plugin. This change was necessary due to issues with the rule causing unintended behavior or conflicts with other rules or coding styles. The decision helps ensure that developers do not encounter unexpected linting errors that could disrupt their workflow.