feat(vue): add `isActive` property to linkable components
Introduced a new `isActive` property in the `useBaseLinkable` composable and integrated it into the `BaseButton` component. This addition enhances the dynamic behavior of linkable components by allowing them to reflect their active state based on the current route. The `isActive` property is computed based on the navigation state which improves UX by providing visual feedback on active links. This change is crucial for components acting as navigational links, ensuring they are visually distinguished when active.