feat(vue): allow string or `InputFiles` as `modelValue` in `useBaseInputFile`
Updated the `useBaseInputFile` composable to accept either a string or an `InputFiles` array as the `modelValue`. This enhancement provides greater flexibility in handling file inputs, accommodating scenarios where a file URL might be used as an initial value. This change simplifies the integration of the composable into environments where file data might already be available in different formats.